# RFID tracking web MVC/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RFID tracking API",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Weapons": {
      "get": {
        "tags": [
          "Weapons"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Weapon"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Weapon"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Weapon"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Weapons"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              }
            }
          }
        }
      }
    },
    "/api/Weapons/{id}": {
      "get": {
        "tags": [
          "Weapons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Weapon"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Weapons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Weapon"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Weapons"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Weapon": {
        "required": [
          "friendlyName",
          "registrationNumber",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "registrationNumber": {
            "minLength": 1,
            "type": "string"
          },
          "friendlyName": {
            "minLength": 1,
            "type": "string"
          },
          "rfidTag": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/WeaponStatus"
          }
        },
        "additionalProperties": false
      },
      "WeaponStatus": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      }
    }
  }
}